backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
gunicorn==21.2.0
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.8
Pillow==10.1.0
bleach==6.1.0
async-timeout==4.0.3

